+++ apps/leaderboard/src/top_arena_server/static/dashboard.js
@@ -9,6 +9,7 @@
   const SIMPLE_AMP_IDS = new Set([
     "blackface63-simple",
     "blackface63-simple-quiet",
+    "blackface63-less-simple",
     "genome-artisan-100-ch1-simple",
     "genome-brit1959-ch1-simple",
     "genome-calibro-normal-simple",